Bug Fix: Bad logic assumed no data in workflow
This addresses an error found by @jackmford in which the SPE bails out early
assuming that there is no longer any data in the workflow.

This commit should fix the issue and it also changes execute to not return a
boolean (denoting data processed) as we don't need to know if operators
produced data, just if there is data in their input queues.